Change project to use local java-time
Previous commits used the java-time library from Clojars. This version
did not support the `As` protocol for `java.time.MonthDay`.

I changed the library source to support `As` for this type and created a
pull request; however, it has not yet been merged and pushed. While I
await this action, I have installed my changes locally and am utilizing
them on this branch.