Change final transformation to activities by day
The previous commit returned a vector of vectors for activities when
executing the last function in `time-calc.adapter`.  However, I really
want to summarize the activities of each day.

This commit changes `time-calc.adapter` to return a map keyed by
activity day with values a vector of all activities (with full start
time) on that day.